requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==23.2.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
deepdiff==6.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
jsonpath==0.82.2
loguru==0.7.2
MarkupSafe==2.1.3
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
paramiko==3.4.0
pluggy==1.3.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
pytest==7.4.4
pytest-rerunfailures==13.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sshtunnel==0.4.0
tomli==2.0.1
urllib3==2.1.0
win32-setctime==1.1.0
